pizza-store/src/requirements.txt
fastapi==0.115.5
pytest==8.3.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
uvicorn==0.32.0
pydantic==2.9.2
pydantic_core==2.23.4
psycopg2==2.9.10

